runtime: mark go-context.S as no-executable-stack and split-stack supported
The .note.GNU-stack section tells the linker that this object does not
require an executable stack.

The .note.GNU-split-stack section tells the linker that functions in
this object can be called directly by split-stack functions, without
require a large stack.

The .note.GNU-no-split-stack section tells the linker that functions
in this object do not have a split-stack prologue.

Change-Id: Ie38235f8b5cafdd2036791a07600ce081ce4b4c3
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/198440
Reviewed-by: Cherry Zhang <cherryyz@google.com>