compiler: fix buglet in function inlining related to sink names
When the compiler writes an inlinable function to the export data,
parameter names are written out (in Export::write_name) using the
Gogo::message_name as opposed to a raw/encoded name. This means that
sink parameters (those named "_") get created with the name "_"
instead of "._" (the name created by the lexer/parser). This confuses
Gogo::is_sink_name, which looks for the latter sequence and not just
"_". This can cause issues later on if an inlinable function is
imported and fed through the rest of the compiler (things that are
sinks are no recognized as such). To fix these issues, change
Gogo::is_sink_name to return true for either variants ("_" or "._").

Fixes golang/go#35586.

Change-Id: I5cddebca0d30932ac76ebd5ca8e765d849323b35
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/207259
Reviewed-by: Ian Lance Taylor <iant@golang.org>