compiler: generate type descriptor for pointer to alias defined in another package
When a type descriptor is needed (for e.g. interface conversion),
if the type is a pointer to a named type defined in another
package, we don't generate the definition of the type descriptor
because it is generated in the package where the type is defined.
However, if the named type is an alias to an unnamed type, its
descriptor is not generated in the other package, and we need to
generate it.

Change-Id: Id54edb1e0df0bf65f755433118da3da00f0186d3
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/210787
Reviewed-by: Ian Lance Taylor <iant@golang.org>