compiler: don't export function type parameter names
When hashing a function type the parameter names are ignored.
Therefore, we should not write them into the export data.
Otherwise, minor changes in the order in which we process the
types can cause the export data to change uselessly, leading to
unnecessary rebuilds.

Change-Id: If3c79474f1ace8a31ee468aa0ec36c846f784aba
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/262818
Trust: Ian Lance Taylor <iant@golang.org>
Reviewed-by: Cherry Zhang <cherryyz@google.com>