compiler: add support for reading embedcfg files
This is the code that parses an embedcfg file, which is a JSON file
created by the go command when it sees go:embed directives.  This code
is not yet called, and does not yet do anything.  It's being sent as a
separate CL to isolate just the JSON parsing code.

Change-Id: I3b13e7434eecc6367bb465e4c29491801bdd24d1
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/281532
Trust: Ian Lance Taylor <iant@golang.org>
Reviewed-by: Than McIntosh <thanm@google.com>