libgo: ensure memmove, memset 8 byte atomicity on ppc64x
Go requires that pointer moves are done 8 bytes at a time,
but gccgo uses libc's memmove and memset which does not require
that, and there are some cases where an 8 byte move might be
done as 4+4.

To enforce 8 byte moves for memmove and memset, this adds a
C implementation in libgo/runtime for memmove and memset to be
used on ppc64le and ppc64. Asm implementations were considered
but discarded to avoid different implementations for different
target ISAs.

Fixes golang/go#41428

Change-Id: I4f3a0969a501054efdcbca8b771a07c006cad9b0
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/294931
Reviewed-by: Ian Lance Taylor <iant@golang.org>
Trust: Lynn Boger <laboger@linux.vnet.ibm.com>