compiler: ignore function type result name in export data
This change ensures that we never output a result name in the export
data if there is only a single result.  Previously we would output a ?
if the single result had a name.  That made the output unstable,
because the hashing ignores the result name, so whether we output a ?
or not depended on how equal hash elements were handled.

For https://gcc.gnu.org/PR104832

Change-Id: I425ff0cdbf44f2987fb5278ca7324b0127523cb4
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/390874
Trust: Ian Lance Taylor <iant@golang.org>
Reviewed-by: Cherry Mui <cherryyz@google.com>