compiler: traverse expressions when exporting constants
When exporting a constant A that is expressed through a constant
B from another package, it is necessary to traverse an expression
representing the constant A to generate a sequence of type casts
from the constant B. Current implementation doesn't collect types
of constants contained in such expressions. This change fetches
these types.

Fixes golang/go#51291

Change-Id: I0f0922a4fa3dd82c18d8cfce7eb83ed98a642505
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/405976
Reviewed-by: Ian Lance Taylor <iant@golang.org>
Reviewed-by: Than McIntosh <thanm@google.com>
Reviewed-by: Ian Lance Taylor <iant@google.com>