compiler: treat S("") as a string constant
The compiler neglected to notice that a conversion from a string
constant to a string type was a valid string constant.

No test case because this only caused a compiler failure when compiling
without optimization, which is not the normal case, and is not a case
that we test.

Fixes golang/go#56113

Change-Id: I99f8f2b91db73b8cb07f415267a9bf2154e1ab54
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/441555
Reviewed-by: Cherry Mui <cherryyz@google.com>
Reviewed-by: Than McIntosh <thanm@google.com>