cmd/go: don't collect package CGOLDFLAGS when using gccgo
They are already collected via cmd/cgo.

The gccgo_link_c test is tweaked to do real linking as with this
change the cgo ldflags are not fully reflected in go build -n output,
since they now only come from the built archive.

This is a backport of https://go.dev/cl/497117 from the main repo.

For golang/go#60287

Change-Id: Ie0bf76b588724265d75dd9a96984dd6cc207f84e
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/511675
Reviewed-by: Than McIntosh <thanm@google.com>
Reviewed-by: Cherry Mui <cherryyz@google.com>