compiler: traverse method declarations
We were not consistently traversing method declarations, which appear
if there is a method without a body.  The gc compiler rejects that
case, but gofrontend currently permits it.  Maybe that should change,
but not today.

This avoids a compiler crash if there are method declarations
with types that require specific functions.  I didn't bother
with a test case because a program with method declarations is
almost certainly invalid anyhow.

Fixes https://gcc.gnu.org/PR117891

Change-Id: I048b35b313629c7a5474cb7df38caa26afafab96
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/633495
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Reviewed-by: Cherry Mui <cherryyz@google.com>